Began adding startup helpers
