+++ ASCOM.Alpaca.Simulators/Startup.cs
@@ -34,12 +34,6 @@ namespace ASCOM.Alpaca.Simulators
             private set;
         }
 
-        internal static string[] Addresses
-        {
-            get;
-            private set;
-        }
-
         // This method gets called by the runtime. Use this method to add services to the container.
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
@@ -47,56 +41,13 @@ namespace ASCOM.Alpaca.Simulators
             services.AddRazorPages();
             services.AddServerSideBlazor();
             services.AddMvc();
+            services.AddBlazoredToast();
 
-            if (ServerSettings.RunSwagger)
-            {
-                services.AddSwaggerGen(c =>
-                {
-                    //This allows for an interesting camera image function but has some issues
-                    //This is off for now
-                    //c.UseOneOfForPolymorphism();
+            var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+            var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
 
-                    c.SwaggerDoc("v1", new OpenApiInfo { Title = $"{ServerSettings.ServerName}", Description = "Please note that the Alpaca API documentation on the ASCOM website is the canonical version. There are several issues with this auto generated version that will be resolved in future versions. This is currently provided only for testing the simulators.", Version = "v0" });
-
-                    var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-                    var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-
-                    if (File.Exists(xmlPath))
-                    {
-                        c.IncludeXmlComments(xmlPath);
-                    }
-                    else
-                    {
-                        Logging.LogInformation("Failed to find simulator xml documentation.");
-                    }
-
-                    if (File.Exists(Path.Combine(AppContext.BaseDirectory, "ASCOM.Common.xml")))
-                    {
-                        xmlPath = Path.Combine(AppContext.BaseDirectory, "ASCOM.Common.xml");
-
-                        c.IncludeXmlComments(xmlPath);
-                    }
-                    else
-                    {
-                        Logging.LogInformation("Failed to find ASCOM.Common xml documentation.");
-                    }
-
-                    if (File.Exists(Path.Combine(AppContext.BaseDirectory, "ASCOM.Alpaca.Razor.xml")))
-                    {
-                        xmlPath = Path.Combine(AppContext.BaseDirectory, "ASCOM.Alpaca.Razor.xml");
-
-                        c.IncludeXmlComments(xmlPath);
-                    }
-                    else
-                    {
-                        Logging.LogInformation("Failed to find ASCOM.Alpaca.Razor xml documentation.");
-                    }
-
-                    c.EnableAnnotations();
-                    c.SchemaFilter<SwaggerExcludeFilter>();
-                    c.MapType<uint>(() => new OpenApiSchema { Type = "integer", Format = "uint32", Minimum = 0, Maximum = 4294967295});
-                });
-            }
+            ASCOM.Alpaca.Razor.StartupHelpers.ConfigureSwagger(services, xmlPath);
+            ASCOM.Alpaca.Razor.StartupHelpers.ConfigureAlpacaAPIBehavoir(services);
 
             // configure basic authentication
             services.AddAuthentication(CookieAuthenticationDefaults.AuthenticationScheme).AddCookie(
@@ -110,47 +61,6 @@ namespace ASCOM.Alpaca.Simulators
             // configure DI for application services
             services.AddScoped<IUserService, UserService>();
             services.AddScoped<AuthorizationFilter>();
-            services.AddBlazoredToast();
-
-            //Do not automatically change JSON to Camel Case
-            services.AddControllers().AddJsonOptions(options => options.JsonSerializerOptions.PropertyNamingPolicy = null);
-
-            //Skip null entries, this should just be the optional driver exception
-            services.AddControllers().AddJsonOptions(options => options.JsonSerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull);
-
-            // Force API Validation errors to return HTTP 400, BadRequest
-            services.Configure<ApiBehaviorOptions>(options =>
-            {
-                options.InvalidModelStateResponseFactory = context =>
-                {
-                    string ErrorMessage = "Unknown Error";
-                    string ResourceMessage = "Unknown Resource";
-
-                    try
-                    {
-                        ErrorMessage = context.ModelState.Values.First(m => m.ValidationState == Microsoft.AspNetCore.Mvc.ModelBinding.ModelValidationState.Invalid).Errors.First().ErrorMessage;
-                    }
-                    catch (Exception ex)
-                    {
-                        Logging.LogError(ex.Message);
-                    }
-
-                    try
-                    {
-                        ResourceMessage = context.HttpContext.Request.Path;
-                    }
-                    catch (Exception ex)
-                    {
-                        Logging.LogError(ex.Message);
-                    }
-
-                    var message = $"A data validation error occurred for request {ResourceMessage} with error message <{ErrorMessage}>. See https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1.";
-
-                    Logging.LogError(message);
-
-                    return new BadRequestObjectResult(message);
-                };
-            });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -178,72 +88,8 @@ namespace ASCOM.Alpaca.Simulators
                                  $"{ServerSettings.ServerName} v1"));
             }
 
-            int port = ServerSettings.ServerPort;
+            Razor.StartupHelpers.ConfigureDiscovery(app);
 
-            //Parse out addresses for the server and start discovery on IPv4 / IPv6 based on which addresses are in use
-            try
-            {
-                var serverAddressesFeature = app.ServerFeatures.Get<IServerAddressesFeature>();
-
-                Addresses = serverAddressesFeature.Addresses.ToArray();
-
-                if (serverAddressesFeature.Addresses.Count > 0)
-                {
-                    var serverAddress = serverAddressesFeature.Addresses.First();
-                    bool localHostOnly = false;
-                    bool ipv6 = false;
-
-                    if (Uri.TryCreate(serverAddress, UriKind.RelativeOrAbsolute, out Uri serverUri))
-                    {
-                        try
-                        {
-                            port = serverUri.Port;
-                            if (serverUri.Host.ToLowerInvariant().Contains("localhost") || IPAddress.IsLoopback(IPAddress.Parse(serverUri.Host)))
-                            {
-                                localHostOnly = true;
-
-                                if (IPAddress.TryParse(serverUri.Host, out IPAddress address))
-                                {
-                                    if (address.AddressFamily == System.Net.Sockets.AddressFamily.InterNetworkV6)
-                                    {
-                                        if (address.IsIPv6LinkLocal)
-                                        {
-                                            localHostOnly = false;
-                                        }
-                                        ipv6 = true;
-                                    }
-                                }
-                            }
-                        }
-                        catch (Exception ex)
-                        {
-                            Logging.LogError(ex.Message);
-                        }
-                    }
-                    else //Invalid Uri, simply parse out port
-                    {
-                        if (serverAddress.Contains(":"))
-                        {
-                            if (int.TryParse(serverAddress.Split(':').Last(), out int result))
-                            {
-                                port = result;
-                            }
-                        }
-
-                        ipv6 = serverAddress.Contains("*") || serverAddress.Contains("+");
-                    }
-
-                    Discovery.DiscoveryManager.Start(port, localHostOnly, ipv6);
-                }
-                else
-                {
-                    Discovery.DiscoveryManager.Start();
-                }
-            }
-            catch (Exception ex)
-            {
-                Logging.LogError(ex.Message);
-            }
 
             //Serve static files, mostly CSS
             app.UseStaticFiles();
@@ -272,7 +118,7 @@ namespace ASCOM.Alpaca.Simulators
                 {
                     if (ServerSettings.AutoStartBrowser) //AutoStart Browser
                     {
-                        Program.StartBrowser(port);
+                        Program.StartBrowser(ServerSettings.ServerPort);
                     }
                 }
                 catch (Exception ex)
